refactor: split zombie system into spawner and HUD
